feat(collapse): code review fixes, move to content folder, bump core version #687
