+++ src/tedi/components/content/collapse/collapse.module.scss
