+++ src/tedi/components/content/collapse/collapse.spec.tsx
