+++ src/tedi/components/layout/footer/footer-section/footer-section.tsx
@@ -4,7 +4,7 @@ import { ReactNode, useContext, useId } from 'react';
 import { Breakpoint, isBreakpointBelow, useBreakpoint } from '../../../../helpers';
 import { Icon } from '../../../base/icon/icon';
 import { Heading } from '../../../base/typography/heading/heading';
-import { Collapse } from '../../../buttons/collapse/collapse';
+import { Collapse } from '../../../content/collapse/collapse';
 import styles from '../footer.module.scss';
 import { FooterContext } from '../footer-context';
